refactor(log): move log entity to domain/log/entities directory
Update all imports to reflect new location of LogEntity
Move and update tests to match new file structure
