+++ src/domain/log/use-cases/__test__/create-log.test.ts
@@ -1,7 +1,7 @@
 import { CreateLog } from "../create-log";
 import { LogRepository } from "@/domain/log/repositories/log.repository";
 import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
-import { LogEntity } from "@/domain/entities/log.entity";
+import { LogEntity } from "@/domain/log/entities/log.entity";
 
 describe("CreateLog UseCase", () => {
   it("delegates to repository.createLog and returns LogEntity", async () => {
