+++ src/domain/log/use-cases/create-log.ts
@@ -1,5 +1,5 @@
 import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
-import { LogEntity } from "@/domain/entities/log.entity";
+import { LogEntity } from "@/domain/log/entities/log.entity";
 import { LogRepository } from "@/domain/log/repositories/log.repository";
 
 interface ICreateLogUseCase {
