+++ src/infrastructure/datasources/__tests__/log.datasource.test.ts
@@ -1,6 +1,6 @@
 import { LogDatasource } from "../log.datasource";
 import { ValidationError } from "@/domain/errors/validation-error";
-import { LogEntity } from "@/domain/entities/log.entity";
+import { LogEntity } from "@/domain/log/entities/log.entity";
 import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 
 // Mock DatabaseClient to return a fake Prisma client
